src/statistics_service/requirements.txt
clickhouse-driver==0.2.7
clickhouse-pool==0.6.0
clickhouse-connect==0.8.16
grpcio==1.71.0
grpcio-reflection==1.71.0
protobuf==5.29.4
confluent-kafka==2.9.0